Handle missing cache dir before pruning temps

+++ src/slic3r/Utils/PresetUpdater.cpp
@@ -264,11 +264,14 @@ bool PresetUpdater::priv::get_file(const std::string &url, const fs::path &targe
 // Remove leftover paritally downloaded files, if any.
 void PresetUpdater::priv::prune_tmps() const
 {
-    for (auto &dir_entry : boost::filesystem::directory_iterator(cache_path))
-		if (is_plain_file(dir_entry) && dir_entry.path().extension() == TMP_EXTENSION) {
-			BOOST_LOG_TRIVIAL(debug) << "Cache prune: " << dir_entry.path().string();
-			fs::remove(dir_entry.path());
-		}
+	if (!fs::exists(cache_path) || !fs::is_directory(cache_path))
+		return;
+
+	for (auto &dir_entry : boost::filesystem::directory_iterator(cache_path))
+			if (is_plain_file(dir_entry) && dir_entry.path().extension() == TMP_EXTENSION) {
+				BOOST_LOG_TRIVIAL(debug) << "Cache prune: " << dir_entry.path().string();
+				fs::remove(dir_entry.path());
+			}
 }
 
 void PresetUpdater::priv::get_missing_resource(const std::string& vendor, const std::string& filename, const std::string& url) const
